services/loader/pyproject.toml
[project]
name = "loader"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.11,<4.0"
dependencies = [
    "pyrecount @ git+https://github.com/dfrll/pyrecount.git",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "pyarrow (>=21.0.0,<22.0.0)",
    "sqlalchemy (>=2.0.41,<3.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "pandas (>=2.3.1,<3.0.0)"
]

[tool.poetry]
packages = [{include = "db", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
